Create a modular structure
